+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-07-13
+### Changed
+- `start_prodwatch()` now requires an `app_name` parameter
+- Manager constructor now requires `app_name` parameter
+- API requests now include app name and process ID where applicable
+
 ## [0.1.0] - 2025-06-24
 ### Added
 - Basic function-level logging.